refactor: change user-home to user-profile and add nav styling
