+++ client/components/index.js
@@ -4,7 +4,7 @@
  * having to figure out which file they belong to!
  */
 export {default as Navbar} from './navbar'
-export {default as UserHome} from './user-home'
+export {default as userprofile} from './user-profile'
 export {Login, Signup} from './auth-form'
 
 export {default as Cart} from './cart'
@@ -12,4 +12,3 @@ export {default as OrderPreview} from './order-preview'
 
 export {default as AllProducts} from './all-products'
 export {default as Record} from './record'
-
