+++ client/components/user-profile.spec.js
@@ -4,19 +4,19 @@ import {expect} from 'chai'
 import React from 'react'
 import enzyme, {shallow} from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
-import {UserHome} from './user-home'
+import {userprofile} from './user-profile'
 
 const adapter = new Adapter()
 enzyme.configure({adapter})
 
-describe('UserHome', () => {
-  let userHome
+describe('userprofile', () => {
+  let userprofile
 
   beforeEach(() => {
-    userHome = shallow(<UserHome email="cody@email.com" />)
+    userprofile = shallow(<userprofile email="cody@email.com" />)
   })
 
   it('renders the email in an h3', () => {
-    expect(userHome.find('h3').text()).to.be.equal('Welcome, cody@email.com')
+    expect(userprofile.find('h3').text()).to.be.equal('Welcome, cody@email.com')
   })
 })
